+++ src/Gui/ConfigDialog.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using RPVoiceChat.Audio;
+using System;
 using System.Collections.Generic;
 using Vintagestory.API.Client;
 using Vintagestory.API.Common;
