+++ src/Gui/SpeechIndicator.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using RPVoiceChat.Audio;
+using System;
 using Vintagestory.API.Client;
 using Vintagestory.API.Common;
